% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_match.R
\name{auto_stratify}
\alias{auto_stratify}
\title{Automatically creates strata for matching}
\usage{
auto_stratify(data, treat, outcome, prog_formula = NULL,
  prog_scores = NULL, size = 2500, held_frac = 0.1,
  held_sample = NULL)
}
\arguments{
\item{data}{data frame with observations as rows, features as columns}

\item{treat}{string giving the name of column designating treatment assignment}

\item{outcome}{string giving the name of column with outcome information}

\item{prog_formula}{formula for building prognostic score model}

\item{size}{numeric, desired size of strata}

\item{held_frac}{numeric between 0 and 1 giving the proportion of samples to be allotted for building the prognostic score}

\item{held_sample}{(optional) a data.frame of held aside samples for building prognostic score model.}
}
\value{
Returns a \code{strata} object
}
\description{
Creates \code{strata} object, given formula for a prognostic model, or pre-calculated prognostic scores
}
