% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{big_match_dopar}
\alias{big_match_dopar}
\title{Big Match Dopar}
\usage{
big_match_dopar(strat, propensity_formula = NULL)
}
\arguments{
\item{strat}{a strata object}

\item{propensity_formula}{a formula for the propensity score.  If
unspecified, all columns except outcome and strata are used.}
}
\value{
a data.frame with pair assignments
}
\description{
Still in development. Match within strata in parallel by calling match_one
with dopar. Doesn't work right now.
}
