% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{big_match_multidplyr}
\alias{big_match_multidplyr}
\title{Big Match v2}
\usage{
big_match_multidplyr(strat, propensity_formula = NULL)
}
\arguments{
\item{strat}{a strata object}

\item{propensity_formula}{the formula for the propensity score}
}
\value{
a data.frame like dat with pair assignments?
}
\description{
Match within strata in parallel by calling match_one
}
