% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{big_match_nstrat}
\alias{big_match_nstrat}
\title{Big Match (not stratified)}
\usage{
big_match_nstrat(strat, propensity_formula = NULL, k = 1)
}
\arguments{
\item{strat}{a strata object}

\item{propensity_formula}{(optional) formula for propensity score}

\item{k}{numeric, the number of control individuals to be matched to each treated individual}
}
\value{
a named factor with matching assignments
}
\description{
For performance testing purposes, match using optmatch but do not stratify
}
