% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_stratify.R
\name{check_base_inputs_auto_stratify}
\alias{check_base_inputs_auto_stratify}
\title{Check inputs from auto_stratify}
\usage{
check_base_inputs_auto_stratify(data, treat, outcome)
}
\arguments{
\item{data}{\code{data.frame} with observations as rows, features as columns}

\item{treat}{string giving the name of column designating treatment
assignment}

\item{outcome}{string giving the name of column with outcome information.
Required if prognostic_scores is specified.  Otherwise it will be inferred
from prog_formula}
}
\value{
nothing; produces errors and warnings if anything is wrong
}
\description{
Not meant to be called externally.  Throws errors if basic auto_stratify 
inputs are incorrect.
}
\keyword{internal}
