% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_stratify.R
\name{check_inputs_auto_stratify}
\alias{check_inputs_auto_stratify}
\title{Check inputs from auto_stratify}
\usage{
check_inputs_auto_stratify(data, treat, outcome, prog_formula, prog_scores,
  size, held_frac, held_sample)
}
\arguments{
\item{data}{\code{data.frame} with observations as rows, features as columns}

\item{treat}{string giving the name of column designating treatment
assignment}

\item{outcome}{string giving the name of column with outcome information.
Required if prog_scores is specified.  Otherwise it will be inferred from
prog_formula}

\item{prog_formula}{formula for building prognostic score model}

\item{prog_scores}{vector of prognostic scores for all individuals in the
data set. Should be in the same order as the rows of \code{data}.}

\item{size}{numeric, desired size of strata (default = 2500)}

\item{held_frac}{numeric between 0 and 1 giving the proportion of controls to
be allotted for building the prognostic score (default = 0.1)}

\item{held_sample}{a data.frame of held aside samples for building prognostic
score model.}
}
\value{
nothing; produces errors and warnings if anything is wrong
}
\description{
Not meant to be called externally.  Throws errors if auto_stratify inputs are
incorrect.
}
