% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_stratify.R
\name{check_outcome}
\alias{check_outcome}
\title{Check Outcome}
\usage{
check_outcome(outcome, data, treat)
}
\arguments{
\item{outcome}{string giving the name of column with outcome information.
Required if prognostic_scores is specified.  Otherwise it will be inferred
from prog_formula}

\item{data}{\code{data.frame} with observations as rows, features as columns}

\item{treat}{string giving the name of column designating treatment
assignment}
}
\value{
nothing
}
\description{
Checks that outcome is a string which is a column in the data
}
