% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_stratify.R
\name{check_prognostic_formula}
\alias{check_prognostic_formula}
\title{Check Prognostic Formula}
\usage{
check_prognostic_formula(prog_formula, data, outcome, treat)
}
\arguments{
\item{prog_formula}{a formula for prognostic score}

\item{data}{\code{data.frame} with observations as rows, features as columns}

\item{outcome}{string giving the name of column with outcome information.
Required if prognostic_scores is specified.  Otherwise it will be inferred
from prog_formula}

\item{treat}{string giving the name of column designating treatment
assignment}
}
\value{
nothing
}
\description{
Check Prognostic Formula
}
