% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{check_prop_formula}
\alias{check_prop_formula}
\title{Check Propensity Formula}
\usage{
check_prop_formula(prop_formula, data, treat)
}
\arguments{
\item{prop_formula}{a formula}

\item{data}{the analysis set data within a stratum}

\item{treat}{the name of the treatment assignment column}
}
\value{
nothing
}
\description{
Check Propensity Formula
}
