% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_stratify.R
\name{check_scores}
\alias{check_scores}
\title{Check Scores}
\usage{
check_scores(prog_scores, data)
}
\arguments{
\item{prog_scores}{vector of prognostic scores for all individuals in the
data set. Should be in the same order as the rows of \code{data}.}

\item{data}{\code{data.frame} with observations as rows, features as columns}
}
\value{
nothing
}
\description{
Checks that prognostic scores are numeric and same length as data
}
