% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_stratify.R
\name{estimate_scores}
\alias{estimate_scores}
\title{Estimate Prognostic Scores}
\usage{
estimate_scores(prognostic_model, analysis_set)
}
\arguments{
\item{prognostic_model}{Model of prognosis}

\item{analysis_set}{data set on which prognostic scores should be estimated}
}
\value{
vector of prognostic scores
}
\description{
Tries to make prognostic scores.  If successfull, returns them, otherwise
throws an error message.  Common failure mode is that the prognostic score is
built on some categorical variable that takes on some values in the analysis
set that are never seen in the pilot set.  Outputs are on the response scale,
(rather than the linear predictor), so the score is the expected value of the
outcome under the control assignement based on the observed covariates.
}
