% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{extract_cut_points.auto_strata}
\alias{extract_cut_points.auto_strata}
\title{Extract cutoffs between strata}
\usage{
\method{extract_cut_points}{auto_strata}(x)
}
\arguments{
\item{x}{an autostrata object}
}
\value{
a vector of the score values delineating cutoffs between strata
}
\description{
Extract cutoffs between strata
}
