% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_stratify.R
\name{fit_prognostic_model}
\alias{fit_prognostic_model}
\title{Fit Prognostic Model}
\usage{
fit_prognostic_model(dat, prognostic_formula, outcome)
}
\arguments{
\item{dat}{data.frame on which model should be fit}

\item{prognostic_formula}{formula for prognostic model}

\item{outcome}{string giving name of column of data where outcomes are
recorded}
}
\value{
a glm or lm object fit from \code{prognostic_formula} on \code{data}
}
\description{
Given a pilot set and a prognostic formula, return the fitted formula.  If
the outcome is binary, fit a logistic regression.  Otherwise, fit a linear
model.
}
