% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_match.R
\name{get_issues}
\alias{get_issues}
\title{Helper for make_issue_table to return issues string}
\usage{
get_issues(row)
}
\arguments{
\item{row}{a row of the data.frame produced in make_issue_table}
}
\value{
Returns a string of potential issues ("none" if everything is fine)
}
\description{
Given a row which summarizes the Treat, Control, Total, and 
Control_Proportion of a stratum, return a string of potential issues with the stratum
}
