% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{get_prop_scores}
\alias{get_prop_scores}
\title{Parse \code{propensity} input to obtain propensity scores}
\usage{
get_prop_scores(propensity, data, treat)
}
\arguments{
\item{propensity}{either a vector of propensity scores, a model for
propensity, or a formula for propensity scores}

\item{data, }{the analysis set data within a stratum}

\item{treat, }{the name of the treatment assignment column}
}
\value{
vector of propensity scores
}
\description{
the \code{propensity} input to \code{plot.auto_strata} or
\code{plot.manual_strata} can be propensity scores, a propensity model, or a
formula for propensity score.  This function figures out which type
\code{propensity} is and returns the propensity scores. Returns the
propensity score on the response scale (rather than the linear predictor), so
the scores are the predited probabilities of treatment.
}
\keyword{internal}
