% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{is.manual_strata}
\alias{is.manual_strata}
\title{Checks \code{manual_strata} class}
\usage{
is.manual_strata(object)
}
\arguments{
\item{object}{any R object}
}
\value{
Returns \code{TRUE} if its argument has class "manual_strata" among its classes and
\code{FALSE} otherwise.
}
\description{
Checks if the target object is a \code{manual_strata} object.
}
