% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{is.strata}
\alias{is.strata}
\title{Checks \code{strata} class}
\usage{
is.strata(object)
}
\arguments{
\item{object}{any R object}
}
\value{
Returns \code{TRUE} if its argument has \code{strata} among its classes and
\code{FALSE} otherwise.
}
\description{
Checks if the target object is a \code{strata} object.
}
\examples{
dat <- make_sample_data()
m.strat <- manual_stratify(dat, treat ~ C1)
is.strata(m.strat) # returns TRUE
}
