% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_stratify.R
\name{is_binary}
\alias{is_binary}
\title{Check if a vector is binary}
\usage{
is_binary(col)
}
\arguments{
\item{col}{a column from a data frame}
}
\value{
logical
}
\description{
return TRUE if the input is logical or if it contains only 0's and 1's
}
\keyword{internal}
