% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{make_ac_plot}
\alias{make_ac_plot}
\title{Make Assignment-Control plot}
\usage{
make_ac_plot(
  x,
  propensity,
  strat,
  strata_lines,
  jitter_prognosis,
  jitter_propensity
)
}
\arguments{
\item{x}{an \code{auto_strata} object returned by \code{\link{auto_stratify}}}

\item{propensity}{ignored unless \code{type = "hist"} or \code{type = "AC"}.
Specifies propensity score information for plots where this is required.
Accepts either a vector of propensity scores, a \code{glm} model for
propensity scores, or a formula for fitting a propensity score model.}

\item{strat}{the number code of the stratum to be plotted. If "all", plots
all strata.}

\item{strata_lines}{default = \code{TRUE}. Ignored unless \code{type = "AC"}.
If TRUE, lines on the plot indicate strata cut points.}

\item{jitter_prognosis}{ignored unless \code{type = "AC"}.  Amount of uniform
random noise to add to prognostic scores in plot.}

\item{jitter_propensity}{ignored unless \code{type = "AC"}.  Amount of
uniform random noise to add to propensity scores in plot.}
}
\description{
Not meant to be called externally.  Helper plot function for \code{strata}
object with \code{type = "AC"}. Produces a Assignment-Control plot of stratum
\code{s}
}
\seealso{
Aikens et al. (preprint) \url{https://arxiv.org/abs/1908.09077} .
  Section 3.2 for an explaination of Assignment-Control plots
}
\keyword{internal}
