% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_stratify.R
\name{make_autostrata_table}
\alias{make_autostrata_table}
\title{Make strata table}
\usage{
make_autostrata_table(qcut)
}
\arguments{
\item{qcut}{the prognostic score quantile cuts}
}
\value{
data.frame of strata definitions
}
\description{
Make strata table
}
\keyword{internal}
