% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{make_distance_matrix}
\alias{make_distance_matrix}
\title{Make Distance Matrix}
\usage{
make_distance_matrix(dat, propensity_model, treat)
}
\arguments{
\item{dat}{a data.frame of observations to be matched}

\item{propensity_model}{a \code{glm} object modeling propensity scores}

\item{treat}{string, the name of the treatment assignment column}
}
\value{
a matrix of distances to be passed to opmatch::pairmatch()
}
\description{
Not meant to be called externally. Makes the distance matrix to be passed to
pairmatch. Similar to \code{match_on.glm} in \code{optmatch} except that the model need not
have been fit on the data we are matching.  Much of this code is copied and 
pasted from the \code{optmatch} package.
}
