% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{make_fm_plot}
\alias{make_fm_plot}
\title{Make Fisher-Mill plot}
\usage{
make_fm_plot(x, propensity, stratum)
}
\arguments{
\item{x}{a \code{strata} object returned by \code{\link{auto_stratify}} or
\code{\link{manual_stratify}}}

\item{propensity}{either a vector if propensity scores, a model for
propensity scores, or a formula for fitting a propensity score model.
Required for "hist" and "FM plots.  Otherwise ignored}

\item{stratum}{the number of the stratum to be plotted.  Required for "hist"
and "FM" plots. Otherwise ignored}
}
\value{
Returns a histogram of propensity scores with strata
}
\description{
Not meant to be called externally.  Helper plot function for \code{strata}
object with type = "FM".  Produces Fisher-Mill plot for individuals within a
stratum.
}
\seealso{
Aikens et al. (preprint) \url{https://arxiv.org/abs/1908.09077} .
  Section 3.2 for an explaination of Fisher-Mill plots
}
