% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_match.R
\name{make_issue_table}
\alias{make_issue_table}
\title{Produce table of Treated and Control by strata}
\usage{
make_issue_table(a_set, treat)
}
\arguments{
\item{a_set}{data frame with observations as rows, features as columns}

\item{treat}{string name of treatment column}
}
\value{
Returns a 3 by [numer of strata] dataframe with Treat, Control, Total, Control Proportion, and Potential Issues
}
\description{
Given a dataframe with strata assigned, tally the number of treated and control samples
}
