% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_stratify.R
\name{make_prog_scores}
\alias{make_prog_scores}
\title{Make Prog Scores}
\usage{
make_prog_scores(prog_model, analysis_set)
}
\arguments{
\item{prog_model}{Model of prognosis}

\item{analysis_set}{data set on which prognostic scores should be estimated}
}
\value{
vector of prognostic scores
}
\description{
Tries to make prognostic scores.  If successfull, returns them, otherwise
throws an error message.  Common failure mode is that the prognostic score is
built on some categorical variable that takes on some values in the analysis
set that are never seen in the pilot set.
}
