% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{make_resid_plot}
\alias{make_resid_plot}
\title{Make Residual Plot}
\usage{
make_resid_plot(x)
}
\arguments{
\item{x}{a \code{strata} object returned by \code{\link{auto_stratify}} or
\code{\link{manual_stratify}}}
}
\value{
Returns the (partial) residual plot(s)
}
\description{
Not yet implemented.  Not meant to be called externally. Helper plot function
for \code{strata} object with type = "residual" Produces partial residual
plots for the prognostic score models
}
