% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sample_data.R
\name{make_sample_data}
\alias{make_sample_data}
\title{Make sample data}
\usage{
make_sample_data(n = 16, p_treat = 0.5)
}
\arguments{
\item{n}{the size of the desired data set}

\item{p_treat}{the probability of treatment}
}
\description{
Makes a simple data frame with treat (binary), outcome (binary), and two
covariates: X1 (categorical) and X2 (continuous). Probability outcome = 1 is
sigmoid(treat + X2)
}
\examples{
  dat <- make_sample_data(n = 50, p_treat = 0.3)
}
