% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{make_SR_plot}
\alias{make_SR_plot}
\title{Make Size-Ratio plot}
\usage{
make_SR_plot(x, label)
}
\arguments{
\item{x}{a \code{strata} object returned by \code{\link{auto_stratify}} or
\code{\link{manual_stratify}}}

\item{label}{ignored unless \code{type = "SR"}. If \code{TRUE}, a clickable
plot is produced. The user may click on any number of strata and press
finish to have those strata labeled.  Note: uses \code{\link{identify}},
which may not be supported on some devices}
}
\description{
Not meant to be called externally.  Helper plot function for \code{strata}.
Produces a scatter plot of strata by size and control proportion.
}
