% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_stratify.R
\name{make_strata_table_qcut}
\alias{make_strata_table_qcut}
\title{Make strata table}
\usage{
make_strata_table_qcut(qcut)
}
\arguments{
\item{qcut}{the prognostic score quantile cuts}
}
\value{
data.frame of strata definitions
}
\description{
Make strata table
}
