% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_functions.R
\name{manual_stratify}
\alias{manual_stratify}
\title{Manual Stratify}
\usage{
manual_stratify(data, treat, covariates, force = FALSE)
}
\arguments{
\item{data}{data.frame with observations as rows, features as columns}

\item{treat}{the name of the column which specifies treatment assignment}

\item{covariates}{a vector strings, which are the names of the columns to be
used as covariates stratification}

\item{force}{a boolean. If true, run even if a variable appears continuous.
(default = FALSE)}
}
\value{
A \code{manual_strata} object
}
\description{
Stratifies a data set based on a set of blocking covariates specified by the
user. Creates a \code{manual_strata} object, which can be passed to
\code{\link{big_match}} for stratified matching or unpacked by the user to be
matched by some other means.
}
\seealso{
\code{\link{auto_stratify}}, \code{\link{new_manual_strata}}
}
