% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{print.auto_strata}
\alias{print.auto_strata}
\title{Print Auto Strata}
\usage{
\method{print}{auto_strata}(strata)
}
\arguments{
\item{strata, }{an \code{auto_strata} object}
}
\description{
Print method for \code{auto_strata} object
}
