% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{print.manual_strata}
\alias{print.manual_strata}
\title{Print Manual Strata}
\usage{
\method{print}{manual_strata}(strata)
}
\arguments{
\item{strata, }{a \code{manual_strata} object}
}
\description{
Print method for \code{manual_strata} object
}
