% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{residual_plot_helper}
\alias{residual_plot_helper}
\title{Helper plot function for \code{strata} object with type = "residual"}
\usage{
residual_plot_helper(auto_strata)
}
\arguments{
\item{auto_strata}{an \code{auto_strata} object}
}
\value{
Returns the (partial) residual plot(s)
}
\description{
Produces partial residual plots for the prognostic scores
}
