% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{scatter_plot_helper}
\alias{scatter_plot_helper}
\title{Helper plot function for \code{strata} object with type = "scatter"}
\usage{
scatter_plot_helper(issue_table, label)
}
\arguments{
\item{strata}{a \code{strata} object}
}
\value{
Returns the scatterplot
}
\description{
Produces a scatterplot of stratum size by control proportion.
}
