% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported_functions.R
\name{strata}
\alias{strata}
\title{Strata function from package Survival}
\usage{
strata(..., na.group = FALSE, shortlabel, sep = ", ")
}
\description{
Strata function from package Survival
}
\keyword{internal}
