% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{summarize_balance}
\alias{summarize_balance}
\title{Calculate balance measures over a dataframe}
\usage{
summarize_balance(data, treat)
}
\arguments{
\item{data, }{a \code{data.frame} from a \code{strata} object}

\item{treat, }{a string giving the name of the treatment assignment column in the dataframe}
}
\value{
Returns a data.frame of balance measures
}
\description{
Calculate Mean over Treat and Control, and difference in Means
}
