% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{summary.strata}
\alias{summary.strata}
\title{Summary for strata object}
\usage{
\method{summary}{strata}(x, ...)
}
\arguments{
\item{x, }{a \code{strata} object}

\item{...}{other arguments}
}
\description{
Summarize important properties of a \code{strata} object
}
\examples{
dat <- make_sample_data()
m.strat <- manual_stratify(dat, treat ~ C1)
summary(m.strat) # returns information about m.strat
}
