% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_match.R
\name{warn_if_continuous}
\alias{warn_if_continuous}
\title{Throws an error if a column is continuous}
\usage{
warn_if_continuous(column, name, force, n)
}
\arguments{
\item{column}{vector or factor column from a data frame}

\item{name}{name of the input column}

\item{force, }{a boolean. If true, warn but do not stop}
}
\value{
Does not return anything
}
\description{
checks if there is a large number of unique values in the input column.
}
