% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_stratify.R
\name{auto_stratify}
\alias{auto_stratify}
\title{Auto Stratify}
\usage{
auto_stratify(data, treat, prognosis, outcome = NULL, size = 2500,
  pilot_fraction = 0.1, pilot_sample = NULL)
}
\arguments{
\item{data}{\code{data.frame} with observations as rows, features as columns}

\item{treat}{string giving the name of column designating treatment
assignment}

\item{prognosis}{information on how to build prognostic scores.  Three
different input types are allowed: \enumerate{ \item vector of prognostic
scores for all individuals in the data set. Should be in the same order as
the rows of \code{data}. \item a \code{formula} for fitting a prognostic
model \item an already-fit prognostic score model}}

\item{outcome}{string giving the name of column with outcome information.
Required if prognostic_scores is specified.  Otherwise it will be inferred from
prog_formula}

\item{size}{numeric, desired size of strata (default = 2500)}

\item{pilot_fraction}{numeric between 0 and 1 giving the proportion of controls to
be allotted for building the prognostic score (default = 0.1)}

\item{pilot_sample}{a data.frame of held aside samples for building prognostic
score model.}
}
\value{
Returns a \code{strata} object
}
\description{
Automatically creates strata for matching based on a prognostic score formula
or a vector of prognostic scores already estimated by the user. Creates a
\code{auto_strata} object, which can be passed to \code{\link{big_match}} for
stratified matching or unpacked by the user to be matched by some other
means.
}
\details{
Stratifying by prognostic score quantiles can be more effective than manually
stratifying a data set because the prognostic score is continuous, thus the
strata produced tend to be of equal size with similar prognosis.

Automatic stratification requires information on how the prognostic scores
should be derived.  This is primarily determined by the specifciation of the
\code{prognosis} argument.  Three main forms of input for \code{prognosis}
are allowed: \enumerate{ \item A vector of prognostic scores. This vector
should be the same length and order of the rows in the data set.  If this
method is used, the \code{outcome} argument must also be specified; this is
simply a string giving the name of the column which contains outcome
information. \item A formula for prognosis (e.g. \code{outcome ~ X1 + X2}).
If this method is used, \code{auto_stratify} will automatically split the
data set into a \code{pilot_set} and an \code{analysis_set}.  The pilot set
will be used to fit a logistic regression model for outcome in the absence of
treatment, and this model will be used to estimate prognostic scores on the
analysis set.  The analysis set will then be stratified based on the
estimated prognostic scores.  In this case the \code{outcome} argument need
not be specified since it can be inferred from the input formula. \item A
model for prognosis (e.g. a \code{glm} object).  If this method is used, the
\code{outcome} argument must also be specified}
}
\section{Troubleshooting}{


  This section suggests fixes for common errors that appear while fitting the
  prognostic score or using it to estimate prognostic scores on the analysis
  set.

  \itemize{

  \item \code{Encountered an error while fitting the prognostic model...
  numeric probabilities 0 or 1 produced}. This error means that the
  prognostic model can perfectly separate positive from negative outcomes.
  Estimating a treatment effect in this case is unwise since an individual's
  baseline characteristics perfectly determine their outcome, regardless of
  whether they recieve the treatment.  This error may also appear on rare
  occaisions when your pilot set is very small (number of observations
  approximately <= number of covariates in the prognostic model), so that
  perfect separation happens by chance.

  \item \code{Encountered an error while estimating prognostic scores ...
  factor X has new levels ... } This may indicate that some value(s) of one
  or more categorical variables appear in the analysis set which were not
  seen in the pilot set. This means that when we try to obtain prognostic
  scores for our analysis set, we run into some new value that our prognostic
  model was not prepared to handle.  There are a few options we have to
  troubleshoot this problem: \itemize{

  \item \strong{Rejection sampling.}  Run \code{auto_stratify} again with the
  same arguments until this error does not occur (i.e. until some
  observations with the missing value are randomly selected into the pilot
  set)

  \item \strong{Eliminate this covariate from the prognostic formula.}

  \item \strong{Remove observations with the rare covariate value from the
  entire data set.} Consider carefully how this exclusion might affect your
  results.

  } }

  Other errors or warnings can occur if the pilot set is too small and the
  prognostic formula is too complicated.  Always make sure that the number of
  observations in the pilot set is large enough that you can confidently fit
  a prognostic model with the number of covariates you want.
}

\seealso{
\code{\link{manual_stratify}}, \code{\link{new_auto_strata}}
}
