% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_match.R
\name{big_match}
\alias{big_match}
\title{Big Match}
\usage{
big_match(strat, propensity_formula = NULL, k = 1)
}
\arguments{
\item{strat}{a strata object}

\item{propensity_formula}{(optional) formula for propensity score}

\item{k}{numeric, the number of control individuals to be matched to each treated individual}
}
\value{
a named factor with matching assignments
}
\description{
Match within strata in series using optmatch
}
