% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_match.R, R/match_functions.R
\name{big_match_dopar}
\alias{big_match_dopar}
\title{Big Match dopar}
\usage{
big_match_dopar(strat, propensity_formula = NULL)

big_match_dopar(strat, propensity_formula = NULL)
}
\arguments{
\item{strat}{a strata object}

\item{strat}{a strata object}
}
\value{
a data.frame like dat with pair assignments?

a data.frame like dat with pair assignments?
}
\description{
Match within strata in parallel by calling match_one with dopar.
Doesn't work right now

Match within strata in parallel by calling match_one with dopar.
Doesn't work right now
}
