% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{big_match_multidplyr}
\alias{big_match_multidplyr}
\title{Big Match Multidplyr}
\usage{
big_match_multidplyr(strat, propensity_formula = NULL, k = 1)
}
\arguments{
\item{strat}{a strata object}

\item{propensity_formula}{the formula for the propensity score}

\item{k}{the number of control individuals to match to each treated
individual}
}
\value{
a data.frame like dat with pair assignments
}
\description{
Still in development. Match within strata in parallel by calling match_one
with multidplyr.  Currently buggy, but returns a result.
}
