% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_match.R
\name{build_prog_model}
\alias{build_prog_model}
\title{Generates prognostic score model from the data}
\usage{
build_prog_model(data, treat, outcome, prog_formula, held_frac = 0.1,
  held_sample = NULL)
}
\arguments{
\item{data}{data frame with observations as rows, features as columns}

\item{treat}{string giving the name of column designating treatment assignment}

\item{outcome}{string giving the name of column with outcome information}

\item{prog_formula}{formula for building prognostic score model}

\item{held_sample, }{(optional) a held aside dataset to be used to fit the prognostic score model}

\item{held_size, }{an integer giving the desired size of the hold-out sample}
}
\value{
Returns list of: m, a \code{glm} object prognostic model, m_set, the model set, and a_set, the analysis set (data - model set)
}
\description{
When the prog_score is not precalculated, subsamples to create a model set (if necessary), then fits a prognostic score on that set.
}
