% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{check_inputs_matcher}
\alias{check_inputs_matcher}
\title{Check inputs to any matching function}
\usage{
check_inputs_matcher(strat, propensity_formula, k)
}
\arguments{
\item{strat}{a strata object}

\item{propensity_formula}{(optional) formula for propensity score}

\item{k}{numeric, the number of control individuals to be matched to each
treated individual}
}
\value{
nothing
}
\description{
Check inputs to any matching function
}
