% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{check_inputs_matcher}
\alias{check_inputs_matcher}
\title{Check inputs to any matching function}
\usage{
check_inputs_matcher(object, propensity, k)
}
\arguments{
\item{object}{a strata object}

\item{propensity}{(optional) formula for propensity score.  If left
blank, all columns of the dataset will be used as covariates (except
outcome, treatment and stratum)}

\item{k}{numeric, the number of control individuals to be matched to each
treated individual}
}
\value{
nothing
}
\description{
Check inputs to any matching function
}
