% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_pilot_set.R
\name{check_pilot_set_options}
\alias{check_pilot_set_options}
\title{Check Pilot set options}
\usage{
check_pilot_set_options(pilot_sample, pilot_fraction, group_by_covariates,
  data)
}
\arguments{
\item{pilot_sample}{a data.frame of held aside samples for building
prognostic score model.}

\item{pilot_fraction}{numeric between 0 and 1 giving the proportion of
controls to be allotted for building the prognostic score (default = 0.1)}

\item{group_by_covariates}{character vector giving the names of covariates to
be grouped by (optional). If specified, the pilot set will be sampled in a
stratified manner, so that the composition of the pilot set reflects the
composition of the whole data set in terms of these covariates.  The
specified covariates must be categorical.}

\item{data}{\code{data.frame} with observations as rows, features as columns}
}
\value{
nothing
}
\description{
Check Pilot set options
}
