% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_stratify.R
\name{check_scores}
\alias{check_scores}
\title{Check Scores}
\usage{
check_scores(prognostic_scores, data)
}
\arguments{
\item{prognostic_scores, }{a numeric vector}

\item{data}{\code{data.frame} with observations as rows, features as columns}
}
\value{
nothing
}
\description{
Checks that prognostic scores are numeric and same length as data
}
