% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_stratify.R
\name{check_scores}
\alias{check_scores}
\title{Check Scores}
\usage{
check_scores(prognostic_scores, data, outcome)
}
\arguments{
\item{prognostic_scores, }{a numeric vector}

\item{data}{\code{data.frame} with observations as rows, features as columns}

\item{outcome}{string giving the name of column with outcome information.
Required if prognostic_scores is specified.  Otherwise it will be inferred
from prog_formula}
}
\value{
nothing
}
\description{
Checks that prognostic scores are the same length as data
}
\keyword{internal}
