% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_issue_table.R
\name{get_issues}
\alias{get_issues}
\title{Get Issues}
\usage{
get_issues(row)
}
\arguments{
\item{row}{a row of the data.frame produced in make_issue_table}
}
\value{
Returns a string of potential issues
}
\description{
Helper for make_issue_table to return issues string.  Given a row which
summarizes the Treat, Control, Total, and Control_Proportion of a stratum,
return a string of potential issues with the stratum.
}
