% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{is.strata}
\alias{is.strata}
\title{Checks strata class}
\usage{
is.strata(object)
}
\arguments{
\item{object}{any R object}
}
\value{
Returns \code{TRUE} if its argument has class "strata" among its classes and
\code{FALSE} otherwise.
}
\description{
Checks if the target object is a \code{strata} object.
}
