% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_match.R
\name{make_distance_matrix}
\alias{make_distance_matrix}
\title{Make distance matrix}
\usage{
make_distance_matrix(dat, propensity_model, treat)
}
\arguments{
\item{dat}{a data.frame of observations to be matched}

\item{propensity_model}{a glm object modeling propensity scores}
}
\value{
a matrix of distances to be passed to pairmatch()
}
\description{
Makes the distance matrix to be passed to pairmatch. Similar to match_on.glm in optmatch
except that the model need not have been fit on the data we are matching
}
