% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{make_hist_plot}
\alias{make_hist_plot}
\title{Make histogram plot}
\usage{
make_hist_plot(auto_strata)
}
\arguments{
\item{auto_strata}{an \code{auto_strata} object}
}
\value{
Returns an arrangement of the column plot and histogram.
}
\description{
Not meant to be called externally.  Helper plot function for \code{strata}
object with type = "hist".  Produces a column plot of prognostic score means
across strata and a histogram of prognostic score, colored by strata.  Not
recommended if there are many strata.
}
