% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{make_hist_plot}
\alias{make_hist_plot}
\title{Make histogram plot}
\usage{
make_hist_plot(x, propensity, strat)
}
\arguments{
\item{x}{a \code{strata} object returned by \code{\link{auto_stratify}} or
\code{\link{manual_stratify}}}

\item{propensity}{ignored unless \code{type = "hist"} or \code{type = "AC"}.
Specifies propensity score information for plots where this is required.
Accepts either a vector of propensity scores, a \code{glm} model for
propensity scores, or a formula for fitting a propensity score model.}

\item{strat}{the number code of the strata to be plotted.  If "all", plots
all strata}
}
\description{
Not meant to be called externally.  Helper plot function for \code{strata}
object with \code{type = "hist"}. Produces a histogram of propensity scores
within a stratum
}
\keyword{internal}
