% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{make_hist_plot}
\alias{make_hist_plot}
\title{Make histogram plot}
\usage{
make_hist_plot(x, propensity, s)
}
\arguments{
\item{x}{a \code{strata} object returned by \code{\link{auto_stratify}} or
\code{\link{manual_stratify}}}

\item{propensity}{either a vector if propensity scores, a model for
propensity scores, or a formula for fitting a propensity score model.
Required for "hist" and "FM" plots.  Otherwise ignored.}

\item{s}{the number code of the strata to be plotted}
}
\value{
Returns a histogram of propensity scores with strata
}
\description{
Not meant to be called externally.  Helper plot function for \code{strata}
object with type = "hist".  Produces a histogram of propensity scores
across strata.
}
