% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{make_resid_plot}
\alias{make_resid_plot}
\title{Make Residual Plot}
\usage{
make_resid_plot(auto_strata)
}
\arguments{
\item{auto_strata}{an \code{auto_strata} object}
}
\value{
Returns the (partial) residual plot(s)
}
\description{
Produces partial residual plots for the prognostic score models
}
\details{
Not yet implemented.  Not meant to be called externally. Helper plot function
for \code{strata} object with type = "residual"
}
