% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sample_data.R
\name{make_sample_data}
\alias{make_sample_data}
\title{Make sample data}
\usage{
make_sample_data(n = 100)
}
\arguments{
\item{n}{the size of the desired data set}
}
\description{
Makes a simple data frame with treat (binary), outcome (binary), and five
covariates: X1 (continuous), X2 (continuous), B1 (binary), B2 (binary), and
C1 (categorical). Probability outcome = 1 is sigmoid(treat + X1). Probability
treatment = 1 is sigmoid(- 0.2 * X1 + X2 - B1 + 2 * B2)
}
\examples{
  # make sample data set of 30 observations
  dat <- make_sample_data(n = 30)
}
