% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{make_scatter_plot}
\alias{make_scatter_plot}
\title{Make scatter plot}
\usage{
make_scatter_plot(x, label)
}
\arguments{
\item{x}{a \code{strata} object returned by \code{\link{auto_stratify}} or
\code{\link{manual_stratify}}}

\item{label}{ignored unless \code{type = "scatter"}. If \code{TRUE}, then
problematic strata are labeled in the scatter plot}
}
\value{
a scatter plot of strata by size and control proportion
}
\description{
Not meant to be called externally.  Helper plot function for \code{strata}.
Checks dependencies and dispatches to \code{\link{make_scatter_plot_gg}} if
able.  Otherwise makes a plot in base R.
}
