% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{make_scatter_plot}
\alias{make_scatter_plot}
\title{Helper plot function for \code{strata} object with type = "scatter"}
\usage{
make_scatter_plot(issue_table, label)
}
\arguments{
\item{strata}{a \code{strata} object}
}
\value{
Returns the scatterplot
}
\description{
Produces a scatterplot of stratum size by control proportion.
}
