% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{make_scatter_plot}
\alias{make_scatter_plot}
\title{Make Scatter Plot}
\usage{
make_scatter_plot(issue_table, label)
}
\arguments{
\item{issue_table}{the \code{issue_table} from a \code{strata} object}

\item{label, }{if true, label problematic strata.  Not recommended if there
are many problematic strata.}
}
\value{
Returns the scatterplot
}
\description{
Not meant to be called externally. Helper plot function for \code{strata}
object with type = "scatter." Generates a scatter plot of all strata by size
and treat:control ratio.  Highlights strata which may be problematic.
}
\seealso{
\code{\link{make_issue_table}}
}
