% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_functions.R
\name{manual_stratify}
\alias{manual_stratify}
\title{Manually creates strata for matching}
\usage{
manual_stratify(data, treat, covariates, force = FALSE)
}
\arguments{
\item{data}{data frame with observations as rows, features as columns}

\item{covariates}{a vector of the columns to be used as covariates for stratification}

\item{force}{a boolean. If true, run even if a variable appears continuous.}
}
\value{
Returns a \code{strata} object
}
\description{
Creates \code{strata} object, given a set of covariates to stratify on
}
