% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_match.R, R/match_functions.R
\name{match_one}
\alias{match_one}
\title{Match One}
\usage{
match_one(dat, propensity_model, treat, k = 1)

match_one(dat, propensity_model, treat, k = 1)
}
\arguments{
\item{dat}{a data.frame with observations as rows, outcome column masked}

\item{dat}{a data.frame with observations as rows, outcome column masked}
}
\value{
a data.frame like dat with pair assignments column

a data.frame like dat with pair assignments column
}
\description{
Match one dataset using the optmatch package

Match one dataset using the optmatch package
}
