% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{plot.strata}
\alias{plot.strata}
\title{Plot method for \code{strata} object}
\usage{
\method{plot}{strata}(strata, type = "scatter", label = FALSE)
}
\arguments{
\item{strata}{a \code{strata} object returned by \code{\link{auto_stratify}}
or \code{\link{manual_stratify}}}

\item{type}{string giving the plot type (default = \code{"scatter"}).  Other
options are \code{"hist"} and \code{"residual"}}

\item{label}{ignored unless \code{type = scatter}. If \code{TRUE}, then
problematic strata are labeled in the scatter plot}
}
\description{
Generates diagnostic plots for the product of a stratification by
\code{\link{auto_stratify}} or \code{\link{manual_stratify}}.  There are
three plot types: \enumerate{ \item "scatter" (default) - produces a scatter
plot of strata by size and treat:control ratio \item "hist" - produces a
histogram of stratum size (not supported for \code{manual strata} objects)
\item "resid" - produces a residual plot for the prognostic model (not
supported for \code{manual strata} objects) }
}
