% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{print.auto_strata}
\alias{print.auto_strata}
\title{Print Auto Strata}
\usage{
\method{print}{auto_strata}(x, ...)
}
\arguments{
\item{x, }{an \code{auto_strata} object}

\item{...}{other arguments}
}
\description{
Print method for \code{auto_strata} object
}
\examples{
dat <- make_sample_data()
a.strat <- auto_stratify(dat, "treat", outcome ~ X1 + X2)
print(a.strat) # prints information about a.strat
}
