% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{print.manual_strata}
\alias{print.manual_strata}
\title{Print Manual Strata}
\usage{
\method{print}{manual_strata}(x, ...)
}
\arguments{
\item{x, }{a \code{manual_strata} object}

\item{...}{other arguments}
}
\description{
Print method for \code{manual_strata} object
}
