% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_functions.R
\name{strata_match_nstrat}
\alias{strata_match_nstrat}
\title{Match without Stratification}
\usage{
strata_match_nstrat(object, propensity = NULL, k = 1)
}
\arguments{
\item{object}{a strata object}

\item{propensity}{(optional) formula for propensity score}

\item{k}{numeric, the number of control individuals to be matched to each
treated individual}
}
\value{
a named factor with matching assignments
}
\description{
Not meant to be called externally.  Match a data set without stratifying.
Used to compare performance with and without stratification.
}
