% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{summarize_balance}
\alias{summarize_balance}
\title{Calculate balance measures over a dataframe}
\usage{
summarize_balance(data, treat, outcome)
}
\arguments{
\item{data, }{a \code{data.frame} from a \code{strata} object}

\item{treat, }{a string giving the name of the treatment assignment column in
the dataframe}

\item{outcome, }{a string giving the name of the outcome column in the
dataframe}
}
\value{
Returns a data.frame of balance measures
}
\description{
Still under development. Calculate mean of each covariate over Treat and
Control groups, and difference in means between groups.
}
