% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{summary.strata}
\alias{summary.strata}
\title{Summary method for strata object}
\usage{
\method{summary}{strata}(object, ...)
}
\arguments{
\item{object}{a \code{strata} object}

\item{...}{other arguments}
}
\description{
Summary method for strata object
}
